[infer] Add type field in Sil.Store
Summary:
It adds typ field in Sil.Store.  The field will be used by the analyzer in the following diffs.

Motivation: Interbo generates a symbolic value when evaluating  expressions including parameter symbols. At that time, it is done with depending on their types, e.g., an integer, a pointer to struct or a pointer to array. Without the type, it is hard to generate a correct symbolic value that will be instantiated later in call sites. Thus, evaluating RHS of the store statement, the type of RHS is better to be given.

Reviewed By: dulmarod

Differential Revision: D17185346

fbshipit-source-id: f0945c40f