Add method to clean capture data
Summary:
There is currently a bug in incremental analysis because the capture data is not reset once the specs files have been invalidated. This has caused a problem where cost issues that should be reported are not spotted. I'm introducing this method so I can use it to fix incremental analysis.

This method is resurrected from D16602417

Reviewed By: ngorogiannis

Differential Revision: D17184401

fbshipit-source-id: e84925324