[nullsafe] Delete nullsafe-gradual tests
Summary:
In D17156724, we forked nullsafe tests, which was a strategy to
introduce nullsafe-gradual mode back then.

The reason was "gradual" mode is a pretty big change in a way Infer
handles annotations, so we wanted to tests both scenarios: gradual and
non-gradual mode.

The plan was to deprecate "non-gradual" tests at some point, hence we
decided to go with duplication.

Now we have a better approach to ensure "gradual" features are well
covered. The approach is the following.

1. [Mostly finished] Improve existings tests so that they cover negative and positive
cases. With this, we can safely add something like
--non-annotated-default UNKNOWN_NULLABILITY to the test config and be sure tests still make
sense (i.e. don't pass simply because annotations don't make sense
anymore)

2. [In progress]. Refactor nullsafe code so that instead of using of Annot.ml everywhere we use a special abstraction telling if the class is annotated with Nullable, Nonnull, or not annotated. With this change, we essenstially have a single place we need to test, which removes the need to have 2 pair of tests for each feature.

3. [To be done]. Introduce Uknown nullability and add small number of tests specifically
for that feature (together with existing tests).

NOTE: I did not rename `nullsafe-default` back to `nullsafe` to not
pollute blame without need.

Reviewed By: artempyanykh

Differential Revision: D17395743

fbshipit-source-id: 3d3e062f6