[inferbo] Address collection add in loop
Summary:
This diff addresses collection adds in loop.  For example,

```
ArrayList<...> a = new ArrayList<>();
for (int i = 0; i < size; i++) {
    a.add(...);
}
// we want to know the size of `a` here!
```

This is a common pattern on initializing a collection in Java.

How we did: Instead of adopting general (but complicated) solutions such as relational domain, we
extended the current alias domain of inferbo, to be able to handle this specific case:

* An array `a` should have size 0, at the entry of the loop.
* The iterating variable `i` should start with 0.
* `add` should be called once inside the loop.

Reviewed By: jvillard

Differential Revision: D17319350

fbshipit-source-id: 99b6acae1