[nullsafe] Don't mention "unannotated" in Models
Summary:
"Unannotated" is misleading and ambiguous concept, it can have different
meanings depending on agreements.

The current logic treats them as Nonnull, which is exactly what we want
to preserve.

(If we need to partially model some functions where we don't have
opinion on some of types in the signature, we can explicitly model
unknown nullability later on).

Note that I am not aiming for substantial refactoring of
modelsTables.ml; the scope of this diff is merely to clarify things.

Reviewed By: ngorogiannis

Differential Revision: D17449347

fbshipit-source-id: 43c798ce7