[cost] Add modeled range
Summary:
`ModeledRange` represents how many times the interval value can be updated by modeled functions.  This
domain is to support the case where there are mismatches between value of a control variable and
actual number of loop iterations.  For example,

```
while((c = file_channel.read(buf)) != -1) { ... }
```

the loop will iterates as the file size, but the control variable `c` does not have that value.  In
these cases, it assigns a symbolic value of the file size to the modeled range of `c`, then which
is used when calculating the overall cost.

Reviewed By: jvillard

Differential Revision: D17476621

fbshipit-source-id: 9a81376e8