[typ] prefer pretty printing to string building
Summary:
My spidey senses were tingling. Next diff uses the `pp` functions
everywhere it was kind of obvious how to change the code to do so. It
doesn't improve perf but is less clowny that way. It might lessen memory
pressure since allocating strings is expensive and this code was doing a
lot of it.

Reviewed By: ngorogiannis

Differential Revision: D17450324

fbshipit-source-id: 632cee584