[nullsafe] Improve readability of the check for constructor initialization
Summary:
It took a while for me to figure out what does this method do; the
reasons were:
1. a lot of names were cryptic and/or misleading
2. because everything is inline one needs to read everything to figure
out what is going on here.

So this diff changes the names a bit and moves some of methods outside.
This is still far not perfect, but I believe it is better than was
before.

Reviewed By: artempyanykh

Differential Revision: D17600175

fbshipit-source-id: ca7175b2e