[required-props] Add more tests showing the ineffectiveness of callee heuristic
Summary: As a heuristic, litho library calls on non-litho callers are not tracked. This is very imprecise and results in FPs and FNs as exemplified by newly added tests. Instead, we should check to see if the summary contains a `build()` method as will be done in the next diff. This diff adds these tests and refactors the test code.

Reviewed By: skcho

Differential Revision: D17809536

fbshipit-source-id: 6dff1868c