[pulse] remember equalities found in branches
Summary:
When we make the decision to go into a branch "v = N" where some
abstract value is compared to a constant, remember the corresponding
equality. This allows to prune simple infeasible paths
intra-procedurally.

Further work is needed to make this useful interprocedurally, for
instance either or both of these ideas could be explored:

- abduce v=N in the precondition and do not apply summaries when the
  equalities in the pre are not satisfied

- prune post-conditions that lead to unsat states where a value has to
  be equal to several different constants

Reviewed By: skcho

Differential Revision: D17906166

fbshipit-source-id: 5cc84abc2