[cost] Conservative array length evaluation
Summary:
This diff adopts an array length evaluation function that is conservative.  It is useful when our
domain cannot express length result precisely.

For example, suppose there is an array pointer `arr_locs` that may point to two arrays `a` and `b`,
and their lengths are `a.length` and `b.length` (symbols), respectively.  Using the usual
evaluation, our current domain cannot express `a.length join b.length` (join of two symbolic
values), so it returns top.

In this case, we can use the conservative function intead.  It evaluates the length as `[0,
a.length.ub + b.length.ub]`, since we know every array length is positive.  The result is not
precise, but better than top.

Reviewed By: ezgicicek

Differential Revision: D17908859

fbshipit-source-id: 7c0b1591b