[annotation reachability] fix bug on android annotation check
Summary:
Java method annotations are ambiguous in that there is no difference between
annotating the return value of a method, and annotating the method itself.
The disambiguation is done entirely based on the meaning of the annotation.

Here, while `UiThread`/`MainThread` are genuine method/class annotations
and not return annotations, the reverse is true for `ForUiThread`/`ForNonUiThread`.
This means that these latter annotations do not determine the thread status of
the method they are attached to.

Here we fix that misunderstanding.

Reviewed By: jvillard

Differential Revision: D17960994

fbshipit-source-id: 5aecfb124