[impurity] Start checking equivalence at materialized addresses in pre
Summary:
Previously, we considered a function which modifies its parameters to be impure even though it might not be modifying the underlying value. This resulted in FPs like the following program in Java:

```
void fresh_pure(int[] a) {
   a = new int[1];
}
```
Similarly, in C++, we considered the following program as impure because it was writing to `s`:
```
Simple* reassign_pure(Simple* s) {
  s = new Simple{2};
  return s;
}
```

This diff fixes that issue by starting the check for address equivalnce in pre-post not directly from the addresses of the stack variables, but from the addresses pointed to by these stack variables. That means, we only consider things to be impure if the actual values pointed by the parameters change.

Reviewed By: skcho

Differential Revision: D18113846

fbshipit-source-id: 3d7c712f3